+++ app/src/main/AndroidManifest.xml
@@ -59,7 +59,7 @@
 
         <!-- [START firebase_service] -->
         <service
-            android:name=".presentation.utils.notification.NotificationService"
+            android:name=".presentation.utils.notification.NotificationFcmService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
