+++ app/src/main/java/com/picone/go4lunch/presentation/ui/fragment/MapsFragment.java
@@ -35,11 +35,11 @@ import java.util.List;
 import java.util.Objects;
 
 import static android.Manifest.permission.ACCESS_FINE_LOCATION;
-import static com.picone.core.data.ConstantParameter.MAPS_CAMERA_ZOOM;
-import static com.picone.core.data.ConstantParameter.MAPS_KEY;
-import static com.picone.core.data.ConstantParameter.REQUEST_CODE;
-import static com.picone.go4lunch.presentation.utils.DailyScheduleHelper.getRestaurantDailyScheduleOnToday;
-import static com.picone.go4lunch.presentation.utils.DailyScheduleHelper.getRestaurantForPlaceId;
+import static com.picone.core.utils.ConstantParameter.MAPS_CAMERA_ZOOM;
+import static com.picone.core.utils.ConstantParameter.MAPS_KEY;
+import static com.picone.core.utils.ConstantParameter.REQUEST_CODE;
+import static com.picone.core.utils.FindInListHelper.getRestaurantDailyScheduleOnToday;
+import static com.picone.core.utils.FindInListHelper.getRestaurantForPlaceId;
 import static com.picone.go4lunch.presentation.utils.GetBitmapFromVectorUtil.getBitmapFromVectorDrawable;
 
 public class MapsFragment extends BaseFragment implements OnMapReadyCallback {
