+++ app/src/main/java/com/picone/go4lunch/presentation/utils/notification/NotificationFcmService.java
@@ -1,0 +1,51 @@
+package com.picone.go4lunch.presentation.utils.notification;
+
+import android.app.AlarmManager;
+import android.app.PendingIntent;
+import android.content.Context;
+import android.content.Intent;
+import android.util.Log;
+
+import androidx.annotation.NonNull;
+
+import com.google.firebase.messaging.FirebaseMessagingService;
+import com.google.firebase.messaging.RemoteMessage;
+import com.picone.go4lunch.R;
+
+import java.util.Calendar;
+import java.util.Date;
+
+import static com.picone.core.utils.ConstantParameter.ALARM_HOUR;
+import static com.picone.core.utils.ConstantParameter.ALARM_MINUTE;
+
+public class NotificationFcmService extends FirebaseMessagingService {
+
+    @Override
+    public void onMessageReceived(RemoteMessage remoteMessage) {
+        if (remoteMessage.getData().size() > 0){
+        setAlarm(this, remoteMessage.getData().get(getString(R.string.intent_extra_name)));}
+    }
+
+    @Override
+    public void onNewToken(@NonNull String token) {
+        Log.d("TAG", "Refreshed token: " + token);
+    }
+
+    private void setAlarm(Context context, String message) {
+        Calendar calendar = Calendar.getInstance();
+        calendar.set(Calendar.HOUR_OF_DAY, ALARM_HOUR);
+        calendar.set(Calendar.MINUTE, ALARM_MINUTE);
+
+        if (calendar.getTime().compareTo(new Date()) < 0)
+            calendar.add(Calendar.DAY_OF_MONTH, 1);
+
+        Intent intent = new Intent(context, NotificationReceiver.class);
+        intent.putExtra(context.getResources().getString(R.string.intent_extra_name), message);
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+        AlarmManager alarmManager = (AlarmManager) context.getSystemService(ALARM_SERVICE);
+
+        if (alarmManager != null) {
+            alarmManager.setExact(AlarmManager.RTC_WAKEUP, calendar.getTimeInMillis(), pendingIntent);
+        }
+    }
+}
