+++ app/src/main/java/com/picone/go4lunch/presentation/utils/notification/NotificationReceiver.java
@@ -1,42 +1,15 @@
 package com.picone.go4lunch.presentation.utils.notification;
 
-import android.app.AlarmManager;
-import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
 import com.picone.go4lunch.R;
 
-import java.util.Calendar;
-import java.util.Date;
-
-import static android.content.Context.ALARM_SERVICE;
-import static com.picone.core.data.ConstantParameter.ALARM_HOUR;
-import static com.picone.core.data.ConstantParameter.ALARM_MINUTE;
-
 public class NotificationReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
         if (intent.getExtras() != null)
             NotificationHelper.createNotification(context, intent.getExtras().getString(context.getResources().getString(R.string.intent_extra_name)));
-    }
-
-    public static void setAlarm(Context context, String message) {
-        Calendar calendar = Calendar.getInstance();
-        calendar.set(Calendar.HOUR_OF_DAY, ALARM_HOUR);
-        calendar.set(Calendar.MINUTE, ALARM_MINUTE);
-
-        if (calendar.getTime().compareTo(new Date()) < 0)
-            calendar.add(Calendar.DAY_OF_MONTH, 1);
-
-        Intent intent = new Intent(context, NotificationReceiver.class);
-        intent.putExtra(context.getResources().getString(R.string.intent_extra_name), message);
-        PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
-        AlarmManager alarmManager = (AlarmManager) context.getSystemService(ALARM_SERVICE);
-
-        if (alarmManager != null) {
-            alarmManager.setExact(AlarmManager.RTC_WAKEUP, calendar.getTimeInMillis(), pendingIntent);
-        }
     }
 }
