+++ app/src/test/java/com/picone/go4lunch/UserViewModelUnitTest.java
@@ -6,7 +6,7 @@ import org.junit.Test;
 
 import java.util.Objects;
 
-import static com.picone.core.data.ConstantParameter.SETTING_START_VALUE;
+import static com.picone.core.utils.ConstantParameter.SETTING_START_VALUE;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
